Move readDatabase()
Move from from createProfile.js to suggestions.js.
